refactor(mobile): split assistant screen components
