+++ apps/mobile/screen/ai-assistant/index.tsx
@@ -1,335 +1,31 @@
-import type {
-  AiAssistantFeedMessage,
-  AiAssistantToolActivity,
-  AiAssistantToolActivityState,
-} from "@services/ai";
 import type { ScrollView as ScrollViewType } from "react-native";
 
-import { IconSymbol } from "@components/IconSymbol";
 import { useAiAssistantChat } from "@hooks/ai/use-ai-assistant-chat";
 import { mapAiAssistantMessagesToFeed } from "@services/ai";
 import { borderWidths, radii, spaceScale, spacingRules } from "@theme/metrics";
-import { fontFaces, fontSizes, fontWeights, lineHeights } from "@theme/typography";
-import { AppCard } from "@ui/primitives/app-card";
 import { AppComposerInput } from "@ui/primitives/app-composer-input";
 import { AppText } from "@ui/primitives/app-text";
 import { Screen } from "@ui/primitives/screen";
-import { Check, SendHorizontal, Sparkles, TriangleAlert } from "lucide-react-native";
-import { Fragment, useCallback, useEffect, useMemo, useRef, useState } from "react";
-import {
-  ActivityIndicator,
-  KeyboardAvoidingView,
-  Platform,
-  Pressable,
-  ScrollView,
-
-  StatusBar,
-  useColorScheme,
-} from "react-native";
-import { useMarkdown } from "react-native-marked";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { KeyboardAvoidingView, Platform, ScrollView, StatusBar } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
-import { useTheme, XStack, YStack } from "tamagui";
+import { XStack, YStack } from "tamagui";
 
-const SUGGESTED_PROMPTS = [
-  { icon: "bike" as const, text: "Tôi đang có chuyến thuê nào không?" },
-  { icon: "calendar" as const, text: "Cho tôi xem đặt chỗ mới nhất." },
-  { icon: "wallet" as const, text: "Ví của tôi hôm nay thế nào?" },
-] as const;
-
-const introMarkdown = `Chào bạn! Mình là **Trợ lý MeBike**.
-
-Mình có thể hỗ trợ các câu hỏi liên quan đến:
-
-- tình trạng thuê xe hiện tại
-- chi tiết đặt chỗ
-- thông tin ví và giao dịch gần đây
-
-Bạn cần giúp gì hôm nay?`;
-
-function getToolColors(state: AiAssistantToolActivityState) {
-  switch (state) {
-    case "done":
-      return {
-        backgroundColor: "$surfaceDefault" as const,
-        borderColor: "$borderSubtle" as const,
-        iconColorKey: "statusSuccess" as const,
-        textTone: "default" as const,
-      };
-    case "error":
-      return {
-        backgroundColor: "$surfaceDefault" as const,
-        borderColor: "$borderSubtle" as const,
-        iconColorKey: "statusDanger" as const,
-        textTone: "danger" as const,
-      };
-    default:
-      return {
-        backgroundColor: "$surfaceDefault" as const,
-        borderColor: "$borderSubtle" as const,
-        iconColorKey: "actionPrimary" as const,
-        textTone: "default" as const,
-      };
-  }
-}
-
-function AssistantMarkdown({ markdown }: { markdown: string }) {
-  const colorScheme = useColorScheme();
-  const theme = useTheme();
-  const elements = useMarkdown(markdown, {
-    colorScheme,
-    styles: {
-      text: {
-        color: theme.textPrimary.val,
-        fontFamily: fontFaces.regular,
-        fontSize: fontSizes.base,
-        fontWeight: fontWeights.regular,
-        lineHeight: lineHeights.base,
-      },
-      paragraph: {
-        marginBottom: spaceScale[1],
-      },
-      strong: {
-        color: theme.textPrimary.val,
-        fontFamily: fontFaces.semibold,
-        fontWeight: fontWeights.semibold,
-      },
-      em: {
-        color: theme.textPrimary.val,
-        fontFamily: fontFaces.medium,
-        fontWeight: fontWeights.medium,
-      },
-      link: {
-        color: theme.textBrand.val,
-        fontFamily: fontFaces.medium,
-        fontWeight: fontWeights.medium,
-      },
-      h1: {
-        color: theme.textPrimary.val,
-        fontFamily: fontFaces.bold,
-        fontSize: fontSizes.xxl,
-        fontWeight: fontWeights.bold,
-        lineHeight: lineHeights.xxl,
-      },
-      h2: {
-        color: theme.textPrimary.val,
-        fontFamily: fontFaces.bold,
-        fontSize: fontSizes.xl,
-        fontWeight: fontWeights.bold,
-        lineHeight: lineHeights.xl,
-      },
-      h3: {
-        color: theme.textPrimary.val,
-        fontFamily: fontFaces.semibold,
-        fontSize: fontSizes.lg,
-        fontWeight: fontWeights.semibold,
-        lineHeight: lineHeights.lg,
-      },
-      li: {
-        color: theme.textPrimary.val,
-        fontFamily: fontFaces.regular,
-        fontSize: fontSizes.base,
-        fontWeight: fontWeights.regular,
-        lineHeight: lineHeights.base,
-      },
-      list: {
-        marginBottom: spaceScale[2],
-      },
-      blockquote: {
-        borderLeftColor: theme.borderFocus.val,
-        borderLeftWidth: 3,
-        marginBottom: spaceScale[2],
-        paddingLeft: spaceScale[3],
-      },
-      codespan: {
-        color: theme.textPrimary.val,
-        fontFamily: fontFaces.medium,
-        fontSize: fontSizes.sm,
-        fontWeight: fontWeights.medium,
-      },
-    },
-  });
-
-  return (
-    <YStack gap="$1">
-      {elements.map((element, index) => <Fragment key={`assistant-markdown-${index}`}>{element}</Fragment>)}
-    </YStack>
-  );
-}
-
-function ToolActivityRow({ activity }: { activity: AiAssistantToolActivity }) {
-  const theme = useTheme();
-  const style = getToolColors(activity.state);
-  const iconColor = theme[style.iconColorKey].val;
-
-  return (
-    <XStack
-      alignItems="center"
-      alignSelf="flex-start"
-      backgroundColor={style.backgroundColor}
-      borderColor={style.borderColor}
-      borderRadius="$4"
-      borderWidth={borderWidths.subtle}
-      gap="$2"
-      paddingHorizontal="$4"
-      paddingVertical="$3"
-    >
-      {activity.state === "running"
-        ? <ActivityIndicator color={iconColor} size="small" />
-        : activity.state === "error"
-          ? <TriangleAlert color={iconColor} size={16} />
-          : <Check color={iconColor} size={16} />}
-
-      <AppText tone={style.textTone} variant="bodySmall">
-        {activity.label}
-      </AppText>
-    </XStack>
-  );
-}
-
-function AssistantBubble({ markdown }: { markdown: string }) {
-  return (
-    <AppCard
-      alignSelf="flex-start"
-      chrome="flat"
-      backgroundColor="$surfaceDefault"
-      borderColor="$borderSubtle"
-      borderTopLeftRadius="$2"
-      borderWidth={borderWidths.subtle}
-      gap="$2"
-      maxWidth="90%"
-      padding="$4"
-      size="default"
-    >
-      <AssistantMarkdown markdown={markdown} />
-    </AppCard>
-  );
-}
-
-function AssistantMessageBlock({ message }: { message: AiAssistantFeedMessage }) {
-  return (
-    <YStack alignItems="flex-start" gap="$2">
-      {message.toolActivities.map((activity, index) => (
-        <YStack key={activity.key} marginTop={index > 0 ? -spaceScale[1] : 0}>
-          <ToolActivityRow activity={activity} />
-        </YStack>
-      ))}
-
-      {message.hasTextContent
-        ? <AssistantBubble markdown={message.markdown} />
-        : null}
-    </YStack>
-  );
-}
-
-function UserMessageBlock({ message }: { message: AiAssistantFeedMessage }) {
-  return (
-    <XStack justifyContent="flex-end">
-      <YStack
-        alignSelf="flex-end"
-        backgroundColor="$actionPrimary"
-        borderRadius="$5"
-        borderTopRightRadius="$2"
-        maxWidth="78%"
-        paddingHorizontal="$5"
-        paddingVertical="$4"
-      >
-        <AppText tone="inverted" variant="bodyStrong">
-          {message.text}
-        </AppText>
-      </YStack>
-    </XStack>
-  );
-}
-
-function TypingIndicator() {
-  return (
-    <AppCard alignSelf="flex-start" borderTopLeftRadius="$2" gap="$2" maxWidth="40%" size="compact">
-      <XStack alignItems="center" gap="$2" minHeight={20}>
-        <YStack backgroundColor="$textDisabled" borderRadius="$round" height={8} opacity={0.8} width={8} />
-        <YStack backgroundColor="$textDisabled" borderRadius="$round" height={8} opacity={0.65} width={8} />
-        <YStack backgroundColor="$textDisabled" borderRadius="$round" height={8} opacity={0.5} width={8} />
-      </XStack>
-    </AppCard>
-  );
-}
-
-function SuggestionChip({
-  disabled,
-  icon,
-  onPress,
-  text,
-}: {
-  disabled: boolean;
-  icon: "bike" | "calendar" | "wallet";
-  onPress: () => void;
-  text: string;
-}) {
-  const theme = useTheme();
-
-  return (
-    <Pressable disabled={disabled} onPress={onPress}>
-      {({ pressed }) => (
-        <XStack
-          alignItems="center"
-          backgroundColor="$surfaceDefault"
-          borderColor="$borderSubtle"
-          borderRadius="$4"
-          borderWidth={borderWidths.subtle}
-          gap="$2"
-          opacity={disabled ? 0.55 : 1}
-          paddingHorizontal="$3"
-          paddingVertical="$3"
-          transform={[{ scale: pressed ? 0.98 : 1 }]}
-        >
-          <IconSymbol color={theme.actionPrimary.val} name={icon} size="caption" />
-          <AppText tone="default" variant="compactStrong">
-            {text}
-          </AppText>
-        </XStack>
-      )}
-    </Pressable>
-  );
-}
-
-function ComposerActionButton({
-  canStop,
-  canSend,
-  onPress,
-  working,
-}: {
-  canStop: boolean;
-  canSend: boolean;
-  onPress: () => void;
-  working: boolean;
-}) {
-  const theme = useTheme();
-  const isDisabled = working ? !canStop : !canSend;
-
-  return (
-    <Pressable disabled={isDisabled} onPress={onPress}>
-      {({ pressed }) => (
-        <XStack
-          alignItems="center"
-          backgroundColor={working || canSend ? "$actionPrimary" : "$surfaceMuted"}
-          borderRadius="$round"
-          height={44}
-          justifyContent="center"
-          opacity={isDisabled ? 0.7 : 1}
-          transform={[{ scale: pressed ? 0.96 : 1 }]}
-          width={44}
-        >
-          {working
-            ? <ActivityIndicator color={theme.onActionPrimary.val} size="small" />
-            : <SendHorizontal color={canSend ? theme.onActionPrimary.val : theme.textDisabled.val} size={18} />}
-        </XStack>
-      )}
-    </Pressable>
-  );
-}
+import {
+  AssistantBubble,
+  AssistantErrorCard,
+  AssistantHeader,
+  AssistantMessageBlock,
+  ComposerActionButton,
+  SuggestionChip,
+  TypingIndicator,
+  UserMessageBlock,
+} from "./components";
+import { INTRO_MARKDOWN, SUGGESTED_PROMPTS } from "./constants";
+import { getAiAssistantErrorMessage } from "./helpers";
 
 export default function AiAssistantScreen() {
   const insets = useSafeAreaInsets();
-  const theme = useTheme();
   const scrollRef = useRef<ScrollViewType | null>(null);
   const [composerText, setComposerText] = useState("");
   const [isSending, setIsSending] = useState(false);
@@ -348,6 +44,7 @@ export default function AiAssistantScreen() {
   const hasConversation = feedMessages.length > 0;
   const canSend = composerText.trim().length > 0;
   const isAssistantWorking = isSending || isBusy || hasRunningToolActivity;
+  const errorMessage = getAiAssistantErrorMessage(error);
   const shouldShowTyping = isAssistantWorking
     && !hasRunningToolActivity
     && feedMessages[feedMessages.length - 1]?.role === "user";
@@ -378,8 +75,7 @@ export default function AiAssistantScreen() {
 
     try {
       await sendTextMessage(trimmedText);
-    }
-    finally {
+    } finally {
       setIsSending(false);
     }
   }, [sendTextMessage]);
@@ -410,37 +106,7 @@ export default function AiAssistantScreen() {
             paddingHorizontal="$5"
             paddingTop={insets.top + spacingRules.hero.paddingTop}
           >
-            <XStack alignItems="center" gap="$3">
-              <XStack
-                alignItems="center"
-                backgroundColor="$overlayGlass"
-                borderRadius="$round"
-                height={52}
-                justifyContent="center"
-                width={52}
-              >
-                <Sparkles color={theme.onSurfaceBrand.val} size={24} />
-              </XStack>
-
-              <YStack flex={1} gap="$1">
-                <AppText tone="inverted" variant="xlTitle">
-                  Trợ lý MeBike
-                </AppText>
-
-                <XStack alignItems="center" gap="$2">
-                  <YStack
-                    backgroundColor="$statusSuccess"
-                    borderRadius="$round"
-                    height={8}
-                    opacity={0.95}
-                    width={8}
-                  />
-                  <AppText opacity={0.92} tone="inverted" variant="bodySmall">
-                    {isAssistantWorking ? "Đang xử lý" : "Đang hoạt động"}
-                  </AppText>
-                </XStack>
-              </YStack>
-            </XStack>
+            <AssistantHeader working={isAssistantWorking} />
           </YStack>
 
           <YStack backgroundColor="$backgroundSubtle" flex={1}>
@@ -458,42 +124,28 @@ export default function AiAssistantScreen() {
               ref={scrollRef}
               showsVerticalScrollIndicator={false}
             >
-              {!hasConversation
-                ? <AssistantBubble markdown={introMarkdown} />
-                : null}
+              {!hasConversation ? <AssistantBubble markdown={INTRO_MARKDOWN} /> : null}
 
               {feedMessages.map(message => (
-                <Fragment key={message.id}>
+                <YStack key={message.id}>
                   {message.role === "user"
                     ? <UserMessageBlock message={message} />
                     : <AssistantMessageBlock message={message} />}
-                </Fragment>
+                </YStack>
               ))}
 
               {shouldShowTyping ? <TypingIndicator /> : null}
-
-              {error
-                ? (
-                    <AppCard alignSelf="flex-start" borderTopLeftRadius="$2" gap="$2" maxWidth="92%" tone="danger">
-                      <AppText tone="danger" variant="label">
-                        Trợ lý tạm thời chưa phản hồi được
-                      </AppText>
-                      <AppText tone="danger" variant="bodySmall">
-                        {error.message}
-                      </AppText>
-                    </AppCard>
-                  )
-                : null}
+              {errorMessage ? <AssistantErrorCard message={errorMessage} /> : null}
             </ScrollView>
 
             <YStack
               backgroundColor="$surfaceDefault"
               borderTopColor="$borderSubtle"
               borderTopWidth={borderWidths.subtle}
-              gap="$4"
+              gap="$3"
               paddingHorizontal="$4"
-              paddingTop="$4"
-              paddingBottom={spaceScale[4] + insets.bottom}
+              paddingTop="$3"
+              paddingBottom={insets.bottom}
             >
               {!hasConversation || feedMessages.length < 3
                 ? (
@@ -522,7 +174,6 @@ export default function AiAssistantScreen() {
               <YStack gap="$1">
                 <AppComposerInput
                   editable={!isAssistantWorking}
-                  // leadingAccessory={<Mic color={theme.textSecondary.val} size={20} />}
                   onChangeText={setComposerText}
                   onSubmitEditing={() => {
                     handleComposerAction();
@@ -540,7 +191,7 @@ export default function AiAssistantScreen() {
                   value={composerText}
                 />
 
-                <AppText align="center" marginTop="$4" tone="subtle" variant="caption">
+                <AppText align="center" marginTop="$2" tone="subtle" variant="caption">
                   Trợ lý AI có thể mắc lỗi. Vui lòng kiểm tra lại thông tin quan trọng.
                 </AppText>
               </YStack>
